split tests, optimize queries for mobile view
