+++ iaso/urls.py
@@ -260,13 +260,13 @@ router.register(r"mobile/stockkeepingunits", StockKeepingUnitMobileViewSet, base
 router.register(r"mobile/stockledgeritems", StockLedgerItemMobileViewSet, basename="mobilestocklegeritems")
 router.register(r"mobile/stockrulesversions", StockRulesVersionMobileViewSet, basename="mobilestockrulesversions")
 
-router.register(r"validation-workflows", ValidationWorkflowViewSet, basename="validationworkflows").register(
+router.register(r"validation-workflows", ValidationWorkflowViewSet, basename="validation_workflows").register(
     r"node-templates",
     ValidationNodeTemplatesView,
-    basename="validation-node-templates",
+    basename="validation_node_templates",
     parents_query_lookups=["workflow__slug"],
 )
-router.register(r"mobile/validation-workflows", ValidationWorkflowMobileViewSet, basename="mobile-validation-workflows")
+router.register(r"mobile/validation-workflows", ValidationWorkflowMobileViewSet, basename="mobile_validation_workflows")
 
 router.registry.extend(plugins_router.registry)
